create separate __init__.py for emitter and move intermediate representation from assign.py to ir.py
