+++ tools/backend/lsq-generator-python/verilog_gen/emitters/__init__.py
@@ -1,0 +1,8 @@
+# verilog_gen/emitters/__init__.py
+from verilog_gen.emitters.emitter import Emitter
+from verilog_gen.emitters.vhdl_emitter import VHDLEmitter
+from verilog_gen.emitters.verilog_emitter import VerilogEmitter
+
+__all__ = [
+    "Emitter", "VHDLEmitter", "VerilogEmitter"
+]
