+++ tools/backend/lsq-generator-python/verilog_gen/operators/__init__.py
@@ -1,5 +1,4 @@
 # verilog_gen/operators/__init__.py
-from verilog_gen.operators.assign import Statement, Val, Bin, Un, WhenElse, Op, BinOp, UnOp, Bit
 from verilog_gen.operators.arithmetic import WrapAdd, WrapAddConst, WrapSub
 from verilog_gen.operators.shifts import CyclicLeftShift
 from verilog_gen.operators.reduction import Reduce
@@ -8,7 +7,6 @@ from verilog_gen.operators.masking import CyclicPriorityMasking
 from verilog_gen.operators.conversions import VecToArray, BitsToOH, BitsToOHSub1, OHToBits
 
 __all__ = [
-    "Statement", "Val", "Bin", "Un", "WhenElse", "Op", "BinOp", "UnOp", "Bit",
     "WrapAdd", "WrapAddConst", "WrapSub",
     "CyclicLeftShift",
     "Reduce",
