+++ tools/backend/lsq-generator-python/verilog_gen/operators/mux.py
@@ -1,8 +1,9 @@
 from verilog_gen.context import Context
-from verilog_gen.emitters.emitter import Emitter
+from verilog_gen.emitters import Emitter
 from verilog_gen.utils import *
 from verilog_gen.signals import *
 from verilog_gen.operators import *
+from verilog_gen.ir import Val, BinOp
 
 
 # ===----------------------------------------------------------------------===#
